Add a "Restriction" section to FinEndo.
This shows how you can restrict the support of a finite endomorphism,
basically constraining the set of elements where it may differ from
the identity. We show how, if the "is this in the support" predicate
is decidable, the restricted map can be shown to be a fin_endo too.