Start working on compositions of substitutions
This chunk supposes that substitutions sigma, tau are given by some
pair of lists. It then defines the correct list for their composition
and shows that the induced map agrees with the composotion on
variables.

It remains to show that it agrees on all terms, but this is the
obvious base case.