Show that the composite of two fin_substs is a fin_subst
This isn't quite as clean as you might hope, because the composition
of the two subst_endos gives an endomorphism on Term, rather than
something that can be a fin_subst. But fin_subst_comp shows that
pre-composing with varTerm gives a fin_subst and comp_subst_is_subst
shows that if you apply subst_endo to this fin_subst, you get back the
original composite map. Phew!