Prove that an injective map on lists implies increasing lengths
This has taken a while, but the point is that if I have a map

  f : A -> B

and know that f is injective on the list aa, taking values in the list
bb, then the set of elements in aa is no bigger than the set of
elements in bb. Writing this in terms of lists, we say that if aa and
bb are both distinct (with no repeats), the length of aa is at most
the length of bb.