The first stage of generalizing compose_fin_endo
I want to compose two fin_mods and conclude that the result is
fin_mod, but it's not actually true (because the first map might have
infinite pre-images).

fin_mod_precomp is the first stage for proving what we actually want
and deals with the problem of pre-images. I think I can lift the proof
of compose_fin_endo to get the rest of the way (and then
compose_fin_endo will be a specialization).