Prove the converse of comp_subst_determined_by_fvs
I think this is supposed to be blindingly obvious, so isn't even
explicitly stated in Eder's paper, but he uses it half-way through the
proof of lemma 2.10.

The core of the statement is that if v is a free variable in t and two
substitutions agree on t then they must agree on v. The proof here
needs decidable equality on F (in order to unpack the fact that two
function terms are equal into the fact that their vectors are
equal). I'm not sure whether there's a clever way to avoid that.

As part of the proof, we also prove a couple of helper lemmas about
converting between vectors and lists. The plan is to forget about all
the vector malarky as soon as we possibly can: everything gets MUCH
easier!

Indeed, I wonder whether we could avoid the decidable equality stuff
above by converting to a sloppier "rose tree" term (that doesn't care
about arities and just has lists). That looks like a lot of typing, so
maybe punt on it for now.