Redefine FiniteProj as a sigma type
Instead of

  Definition FiniteProj (A B : Type) (p : A -> B) : Prop :=
    exists l : list A, FullProj p l.

we now have

  Definition FiniteProj (A B : Type) (p : A -> B) : Type :=
    { l : list A | FullProj p l }.

The point is that computations can now destruct a FiniteProj to get
the list witnessing its finiteness.

The main use is in fin_mod. Now, if you have an object of
type (fin_mod f), then a computation can extract the finite list of
domain elements.

The biggest resulting changes are in Perm.v, where we can now actually
define the inverse map (before, we had to destruct the exist Prop, so
could only assert that the inverse map existed.