Redefine fin_subst and var_subst as sigma types
A fin_subst is now a finite substitution, rather than a proof that
some given substitution is finite. Similarly, a var_subst is a finite
endomorphism on V, rather than a proof that a given endomorphism is
finite.

This causes a certain amount of plumbing, but some of the lemmas end
up with better types and (the main point) the sections at the end of
Generality.v now have a less ridiculous set of hypotheses.

This *still* isn't quite enough to state lemma 2.10 cleanly: there are
a few too many fin_subst_subst calls floating around. But it's
definitely getting close.