Make an informative version of extends1/extends2 in InjMod
Now we have is_ext1, where is_ext1 l0 l1 a means "l0 extends to l1 by
adding a". The definition of extends is unchanged, but now we can
define get_extension, which gets the thing to be added and (after a
bit of elimination gymnastics!) show is_ext1_get_extension and
is_ext2_get_extension, which something using the theory can use as
part of its inductive step.